requirements.txt
fireworks_ai==0.14.0
groq==0.9.0
mistralai==0.4.1
octoai==1.4.0
openai==1.35.9
protobuf==5.27.2
rich==13.7.1
google-generativeai
anthropic
reka-api
pynput


